add new configurable websocket api
