+++ jupyter_server/serverapp.py
@@ -26,36 +26,7 @@ import warnings
 import webbrowser
 from base64 import encodebytes
 
-try:
-    import resource
-except ImportError:
-    # Windows
-    resource = None  # type:ignore[assignment]
-
-from jinja2 import Environment, FileSystemLoader
-from jupyter_core.paths import secure_write
-
-from jupyter_server.services.kernels.handlers import ZMQChannelsHandler
-from jupyter_server.transutils import _i18n, trans
-from jupyter_server.utils import ensure_async, pathname2url, urljoin
-
-# the minimum viable tornado version: needs to be kept in sync with setup.py
-MIN_TORNADO = (6, 1, 0)
-
-try:
-    import tornado
-
-    assert tornado.version_info >= MIN_TORNADO
-except (ImportError, AttributeError, AssertionError) as e:  # pragma: no cover
-    raise ImportError(_i18n("The Jupyter Server requires tornado >=%s.%s.%s") % MIN_TORNADO) from e
-
-from tornado import httpserver, ioloop, web
-from tornado.httputil import url_concat
-from tornado.log import LogFormatter, access_log, app_log, gen_log
-
-if not sys.platform.startswith("win"):
-    from tornado.netutil import bind_unix_socket
-
+from jupyter_client import KernelManager
 from jupyter_client.kernelspec import KernelSpecManager
 from jupyter_client.manager import KernelManager
 from jupyter_client.session import Session
@@ -63,6 +34,9 @@ from jupyter_core.application import JupyterApp, base_aliases, base_flags
 from jupyter_core.paths import jupyter_runtime_dir
 from jupyter_events.logger import EventLogger
 from nbformat.sign import NotebookNotary
+from tornado import httpserver, ioloop, web
+from tornado.httputil import url_concat
+from tornado.log import LogFormatter, access_log, app_log, gen_log
 from traitlets import (
     Any,
     Bool,
@@ -122,16 +96,27 @@ from jupyter_server.services.contents.filemanager import (
     AsyncFileContentsManager,
     FileContentsManager,
 )
-from jupyter_server.services.contents.largefilemanager import AsyncLargeFileManager
+from jupyter_server.services.contents.largefilemanager import (
+    AsyncLargeFileManager,
+    LargeFileManager,
+)
 from jupyter_server.services.contents.manager import (
     AsyncContentsManager,
     ContentsManager,
 )
+from jupyter_server.services.kernels.connection.base import (
+    BaseKernelWebsocketConnection,
+)
+from jupyter_server.services.kernels.connection.channels import (
+    ZMQChannelsWebsocketConnection,
+)
+from jupyter_server.services.kernels.kernel_broker import KernelWebsocketBroker
 from jupyter_server.services.kernels.kernelmanager import (
     AsyncMappingKernelManager,
     MappingKernelManager,
 )
 from jupyter_server.services.sessions.sessionmanager import SessionManager
+from jupyter_server.traittypes import TypeFromClasses
 from jupyter_server.utils import (
     check_pid,
     fetch,
@@ -140,6 +125,55 @@ from jupyter_server.utils import (
     url_path_join,
     urlencode_unix_socket_path,
 )
+
+try:
+    import resource
+except ImportError:
+    # Windows
+    resource = None  # type:ignore[assignment]
+
+from jinja2 import Environment, FileSystemLoader
+from jupyter_core.paths import secure_write
+
+from jupyter_server.transutils import _i18n, trans
+from jupyter_server.utils import ensure_async, pathname2url, urljoin
+
+# the minimum viable tornado version: needs to be kept in sync with setup.py
+MIN_TORNADO = (6, 1, 0)
+
+try:
+    import tornado
+
+    assert tornado.version_info >= MIN_TORNADO
+except (ImportError, AttributeError, AssertionError) as e:  # pragma: no cover
+    raise ImportError(_i18n("The Jupyter Server requires tornado >=%s.%s.%s") % MIN_TORNADO) from e
+
+
+if not sys.platform.startswith("win"):
+    from tornado.netutil import bind_unix_socket
+
+
+try:
+    import resource
+except ImportError:
+    # Windows
+    resource = None  # type:ignore[assignment]
+
+
+# the minimum viable tornado version: needs to be kept in sync with setup.py
+MIN_TORNADO = (6, 1, 0)
+
+try:
+    import tornado
+
+    assert tornado.version_info >= MIN_TORNADO
+except (ImportError, AttributeError, AssertionError) as e:  # pragma: no cover
+    raise ImportError(_i18n("The Jupyter Server requires tornado >=%s.%s.%s") % MIN_TORNADO) from e
+
+
+if not sys.platform.startswith("win"):
+    from tornado.netutil import bind_unix_socket
+
 
 # -----------------------------------------------------------------------------
 # Module globals
@@ -157,7 +191,10 @@ JUPYTER_SERVICE_HANDLERS = dict(
     config=["jupyter_server.services.config.handlers"],
     contents=["jupyter_server.services.contents.handlers"],
     files=["jupyter_server.files.handlers"],
-    kernels=["jupyter_server.services.kernels.handlers"],
+    kernels=[
+        "jupyter_server.services.kernels.handlers",
+        "jupyter_server.services.kernels.websocket",
+    ],
     kernelspecs=[
         "jupyter_server.kernelspecs.handlers",
         "jupyter_server.services.kernelspecs.handlers",
@@ -224,6 +261,7 @@ class ServerWebApplication(web.Application):
         *,
         authorizer=None,
         identity_provider=None,
+        kernel_websocket_connection_class=None,
     ):
         if identity_provider is None:
             warnings.warn(
@@ -259,6 +297,7 @@ class ServerWebApplication(web.Application):
             jinja_env_options,
             authorizer=authorizer,
             identity_provider=identity_provider,
+            kernel_websocket_connection_class=kernel_websocket_connection_class,
         )
         handlers = self.init_handlers(default_services, settings)
 
@@ -282,6 +321,7 @@ class ServerWebApplication(web.Application):
         *,
         authorizer=None,
         identity_provider=None,
+        kernel_websocket_connection_class=None,
     ):
 
         _template_path = settings_overrides.get(
@@ -362,6 +402,7 @@ class ServerWebApplication(web.Application):
             authorizer=authorizer,
             identity_provider=identity_provider,
             event_logger=event_logger,
+            kernel_websocket_connection_class=kernel_websocket_connection_class,
             # handlers
             extra_services=extra_services,
             # Jupyter stuff
@@ -1462,6 +1503,13 @@ class ServerApp(JupyterApp):
             return "jupyter_server.gateway.managers.GatewaySessionManager"
         return SessionManager
 
+    kernel_websocket_connection_class = Type(
+        default_value=ZMQChannelsWebsocketConnection,
+        klass=BaseKernelWebsocketConnection,
+        config=True,
+        help=_i18n("The kernel websocket connection class to use."),
+    )
+
     config_manager_class = Type(
         default_value=ConfigManager,
         config=True,
@@ -2025,6 +2073,7 @@ class ServerApp(JupyterApp):
             self.jinja_environment_options,
             authorizer=self.authorizer,
             identity_provider=self.identity_provider,
+            kernel_websocket_connection_class=self.kernel_websocket_connection_class,
         )
         if self.certfile:
             self.ssl_options["certfile"] = self.certfile
@@ -2820,7 +2869,7 @@ class ServerApp(JupyterApp):
         self.remove_browser_open_files()
         await self.cleanup_extensions()
         await self.cleanup_kernels()
-        await ZMQChannelsHandler.close_all()
+        await self.kernel_websocket_connection_class.close_all()
         if getattr(self, "kernel_manager", None):
             self.kernel_manager.__del__()
         if getattr(self, "session_manager", None):
